# 12 sql type delta - bigint

+++ migrations/01_metric.sql
@@ -2,6 +2,6 @@ CREATE TABLE IF NOT EXISTS metric
 (
     id   VARCHAR(50) PRIMARY KEY,
     type VARCHAR(50),
-    delta integer,
+    delta bigint,
     value decimal
 );
